Bench: 25255797

+++ Halogen/src/Network.cpp
@@ -73,7 +73,7 @@ int16_t Network::QuickEval() const
     int32_t output = outputBias * PRECISION;
 
     for (size_t i = 0; i < HIDDEN_NEURONS; i++)
-        output += std::max(int32_t(0), Zeta.back()[i]) * outputWeights[i];
+        output += std::max(int16_t(0), Zeta.back()[i]) * outputWeights[i];
 
     return output / SQUARE_PRECISION;
 }
